// src/app/_core/Models/Film.ts
export class Film {
    maPhim: number = 0;
    tenPhim: string = '';
    biDanh: string = '';
    trailer: string = '';
    hinhAnh: string = '';
    moTa: string = '';
    maNhom: string = '';
    ngayKhoiChieu: Date = new Date();
    danhGia: number = 0;
    hot: boolean = true;
    dangChieu: boolean = true;
    sapChieu: boolean = true;
    constructor() {

    }

}


// {
//     "maPhim": 4956;
//     "tenPhim": "Wanted2";
//     "biDanh": "wanted2";
//     "trailer": "https://www.youtube.com/watch?v=W7PfG4iweZk ";
//     "hinhAnh": "https://movienew.cybersoft.edu.vn/hinhanh/wanted_gp01.jpg";
//     "moTa": "wanted is a 2008 American action thriller directed by Tinur Beknambetov and wirtten by Michael Brandt"
//     "maNhom": "GP01";
//     "ngayKhoiChieu": "2021-09-27T00:00:00";
//     "danhGia": 10;
//     "hot": true;
//     "dangChieu": true;
//     "sapChieu": true;
// }